// frontend/src/types/event.ts
export type Event = {
    id: string;
    name: string;
    code: string | null;
    poblacion: string | null;
    lugar: string | null;
    reference: string | null;
    notes: string | null;
    startDate: string | null;
    endDate: string | null;
    horaMontaje: string | null;
    horaPrueba: string | null;
    horaComienzo: string | null;
    horaFin: string | null;
    horaDesmontaje: string | null;
    createdAt: string;
    updatedAt: string;
};